Used server socket for creating servers removing the threading
