+++ server_client/clientsocket.py
@@ -1,16 +1,42 @@
 import socket
 
 host, port = "localhost", 9999
-data = input("Enter message to send: ")
 
-# Create a socket (SOCK_STREAM means a TCP socket)
-with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
-    # Connect to server and send data
-    sock.connect((host, port))
-    sock.send(data.encode("utf-8"))
+s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+print('Socket Made')
 
-    # Receive data from the server and shut down
-    received = str(sock.recv(1024), "utf-8")
+s.connect((host, port))
 
-print("Sent:     {}".format(data))
-print("Received: {}".format(received))
+
+def mode0():
+    s.send('0'.encode('utf-8'))
+    while True:
+        message = input('What would you like to say \n')
+        if message == 'break':
+            s.send('break'.encode('utf-8'))
+            break
+        s.send(message.encode('utf-8'))
+
+        if (s.recv(1024)).decode('utf-8'):
+            print('Message Worked')
+
+
+def mode1():
+    s.send('1'.encode('utf-8'))
+    while True:
+        message = input('What would you like to guess \n')
+        if message == 'break':
+            s.send('break'.encode('utf-8'))
+            break
+        s.send(message.encode('utf-8'))
+
+        if (s.recv(1024)).decode('utf-8') == 'Won':
+            print('You won')
+            break
+
+
+choice = input('Mode?\n')
+if choice == '0':
+    mode0()
+if choice == '1':
+    mode1()
